pyproject.toml
[tool.poetry]
name = "curr-bot-v2"
version = "0.1.0"
description = ""
authors = ["mama tvoia"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pika = "^1.3.2"
python-dotenv = "^1.0.0"
python-telegram-bot = "^20.5"
web3 = "^6.5.0"
python-binance = "^1.0.19"
rpyc = "^5.3.1"
jinja2 = "^3.1.2"
redis = "^5.0.1"
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

